Better Registration form
This integrates the updates to the Registration page from Deep, and includes my edits to make it connect to the database. There was an issue where input fields were not completed. So code from the old registration form that did work was substituted. we have an issue where the input field script lifts up and out of the form.